server/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.15
numpy==1.15.4
protobuf==3.7.1
requests==2.21.0
six==1.12.0
soupsieve==1.9
SQLAlchemy==1.3.2
urllib3==1.24.1
virtualenv==16.2.0
Werkzeug==0.15.2
